Combined HMDs and Controllers into a single client.
No need to run one binary to track HMDs and another to track controllers.
This code can track both at the same time and write the output to different
FreePIE slots. It can also read yaw/pitch/roll data from a specific FreePIE
slot and combine it with the positional tracking for the HMD.